Make panel toolbars always visible
Elements that change their visibility state based on the mouse hover on a panel draw attention as the user moves their cursor around the app or into a panel. They also make it harder to see where to move your cursor within a panel to use the actions on the toolbar.

By always showing the toolbar the user knows where the control icons are. The other elements do not shift when the toolbar appears/disappears since it is always visible.